First draft of blakcomp compiling optimizations.
Work in progress, attempting to improve the performance of kod by having
the compiler optimize the code during compilation. The current compiler
only does constant folding, I have added discarding $ assignments to $
local vars and two optimizations on the Nth() C call.

The compiler will also notify if a Send() call is duplicated inside a
message (but does not modify, as there are valid reasons to call the
same message multiple times). 'If' conditions that evaluate to a
constant are also flagged but not modified.

Optimizations are output to console and logged in \kod\optimize.txt.
Included the log file here for a current dev branch build.